add texture support to renderer2d

+++ Hazel/src/Hazel/Renderer/Buffer.h
@@ -57,7 +57,7 @@ namespace Hazel {
         std::string Name;
         ShaderDataType Type;
         uint32_t Size;
-        uint32_t Offset;
+        size_t Offset;
         bool Normalized;
 
         BufferElement() = default;
@@ -130,7 +130,7 @@ namespace Hazel {
 
     private:
         void CalculateOffsetsAndStride() {
-            uint32_t offset = 0;
+            size_t offset = 0;
             m_Stride = 0;
             for (auto &element : m_Elements) {
                 element.Offset = offset;
